# openapi.yaml
openapi: 3.0.3
info:
  title: Baseera API
  version: v1
  description: |
    **FinTech Hackathon MVP** — Personal Finance Management with AI-Driven Subscription Detection.

    Baseera helps users track spending, detect recurring subscriptions automatically,
    and receive actionable financial health insights via a single dashboard.

    ### Authentication
    All endpoints (except Register & Login) require a valid JWT Bearer token
    in the `Authorization` header.

    ### Rate Limiting
    - **Auth endpoints**: 5 requests/minute per IP (`IpRateLimit`).
    - **Protected endpoints**: 40 requests/minute per user (`UserRateLimit`).
    - Exceeding limits returns `429 Too Many Requests` with a `Retry-After` header.
  contact:
    name: Baseera Team
    email: team@baseera.com
  license:
    name: MIT
servers:
  - url: https://baseera.runasp.net
    description: Production Server
  - url: https://localhost:7181
    description: Local Development (HTTPS)
  - url: http://localhost:5002
    description: Local Development (HTTP)
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/Auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new user account and returns a JWT token (24 h expiry).
        Also serves as onboarding — sets the user's monthly income for budget calculations.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: demo@baseera.com
              password: "P@ssw0rd!"
              firstName: Ahmed
              lastName: Hassan
      responses:
        "200":
          description: Registration successful — JWT token returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                token: "eyJhbGciOiJIUzI1NiIs..."
                expiration: "2026-04-30T12:00:00Z"
                userId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                email: demo@baseera.com
                fullName: Ahmed Hassan
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
                message: A user with this email already exists.
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/Auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Authenticates the user and returns a JWT token valid for 24 hours.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: demo@baseera.com
              password: "P@ssw0rd!"
      responses:
        "200":
          description: Login successful — JWT token returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
                message: Invalid email or password.
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  # ═══════════════════════════════ ACCOUNTS ════════════════════════════════
...
  /api/Subscriptions:
    get:
      tags: [Subscriptions]
      summary: Get all subscriptions
      description: |
        Returns every AI-detected subscription for the authenticated user.
        Automatically evaluates at-risk status before returning results.
      responses:
        "200":
          description: A list of subscriptions with current status.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
              example:
                - id: "e0b4f5a6-7c89-0123-defa-456789012efa"
                  serviceName: Netflix
                  monthlyCost: 49.99
                  lastPaymentDate: "2026-04-15T00:00:00Z"
                  lastActivityDate: "2026-04-20T14:30:00Z"
                  usageScore: 0.85
                  status: Active
                - id: "f1c5a6b7-8d90-1234-efab-567890123fab"
                  serviceName: Gym Membership
                  monthlyCost: 199.00
                  lastPaymentDate: "2026-04-01T00:00:00Z"
                  lastActivityDate: "2026-02-10T09:00:00Z"
                  usageScore: 0.15
                  status: AtRisk
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/Subscriptions/{id}/cancel:
    patch:
      tags: [Subscriptions]
      summary: Cancel a subscription
      description: Sets the subscription status to **Cancelled**.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Subscription ID.
      responses:
        "200":
          description: Subscription cancelled successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Subscription is already cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: Subscription is already cancelled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Subscription not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: Subscription not found.
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"

  # ═══════════════════════════════ DASHBOARD ══════════════════════════════
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste the token returned by `/api/Auth/login` or `/api/Auth/register`.

  # ═══════════════════════════════ SCHEMAS ═════════════════════════════════
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          example: demo@baseera.com
        password:
          type: string
          format: password
          minLength: 6
          description: "Must contain: uppercase, lowercase, digit, and special character."
          example: "P@ssw0rd!"
        firstName:
          type: string
          maxLength: 100
          example: Ahmed
        lastName:
          type: string
          maxLength: 100
          example: Hassan
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: demo@baseera.com
        password:
          type: string
          format: password
          example: "P@ssw0rd!"
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Bearer token (24 h expiry).
        expiration:
          type: string
          format: date-time
        userId:
          type: string
          description: ASP.NET Identity user ID.
        email:
          type: string
          format: email
        fullName:
          type: string

    # ── Accounts ──────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceName:
          type: string
          example: Netflix
        monthlyCost:
          type: number
          format: decimal
          example: 49.99
        lastPaymentDate:
          type: string
          format: date-time
        lastActivityDate:
          type: string
          format: date-time
          nullable: true
        usageScore:
          type: number
          format: decimal
          minimum: 0
          maximum: 1
          description: "AI-computed score (0 = unused, 1 = heavily used)."
          example: 0.85
        status:
          type: string
          enum: [Active, AtRisk, Cancelled]

    # ── Dashboard ─────────────────────────────────────────────────────────
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
    GlobalError:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string

  # ═══════════════════════════════ RESPONSES ═══════════════════════════════
  responses:
    BadRequest:
      description: Validation failed or malformed request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
          example:
            errors:
              - "The Email field is required."
              - "The Password field must be at least 6 characters."
    Unauthorized:
      description: Missing or invalid JWT token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GlobalError"
          example:
            status: 401
            message: Unauthorized access.
            timestamp: "2026-04-29T12:00:00Z"
    Forbidden:
      description: Authenticated but not authorized to access this resource.
    TooManyRequests:
      description: Rate limit exceeded — check the `Retry-After` header.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the next request is allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            error: "Too many requests. Please check the Retry-After header."
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GlobalError"
          example:
            status: 500
            message: An unexpected error occurred. Please try again later.
            timestamp: "2026-04-29T12:00:00Z"